Fixed assorted compiler warnings that showed up switching between LLVM and GCC.
git-svn-id: https://crag.googlecode.com/svn/trunk@442 03cb304c-929a-b8d2-5531-aad0e9782889

+++ src/crag/gfx/Renderer.cpp
@@ -333,7 +333,7 @@ void gfx::Renderer::Render()
 	// update fps
 	memmove(_fps_history, _fps_history + 1, sizeof(* _fps_history) * (_fps_history_size - 1));
 	_fps_history[_fps_history_size - 1] = sys::GetTime();
-	STAT_SET (fps, double(_fps_history_size - 1) / (_fps_history[_fps_history_size - 1] - _fps_history[0]));
+	STAT_SET (fps, float(_fps_history_size - 1) / float(_fps_history[_fps_history_size - 1] - _fps_history[0]));
 #endif
 	
 	// Regulator feedback.
